+++ NEWS.md
@@ -1,5 +1,7 @@
 ## nestedLogit 0.3.5
 
+This is a major release of the package adding considerable functionality for plotting and other features
+
 * add gators data, food choice of alligators from Agresti
 * added an article illustrating use of the `ggeffects` package for nestedLogit models [only an article, because work on this is continuing]
 * added an article showing use of the `easystats` packages for nestedLogit models
@@ -8,6 +10,7 @@
 * Better description of nested logit and other models in README
 * changed default colors for `plot.nestedLogit()` to use scales::hue_pal() for consistency with ggplot
 * added `scale` argument to `plot.nestedLogit()` to allow plotting on the logit (log-odds) scale (`scale = "logit"`) in addition to the default probability scale (`scale = "prob"`)
+* added `as.tree()` method for ASCII printing of the tree of nested dichotomies
 
 ## nestedLogit 0.3.4